broke out interpreter/python bin methods
